Fixed RUM Speed Index calculation on images
When srcset is used on images the src attribute is ignored and the browser renders based on the viewport size. The currentSrc attribute is the right attribute to check in such cases.
How does this impact RUM Speed Index?
When calculating speedindex the algorithm right now relies on src attribute, however what is rendered on the screen is a totally different image src. Hence the when the time of render is retrieved for the image, it return 0 causing a miscalculation in speed index.